moved the profile sidepublication to its parent component landingpage
