+++ src/components/LandingPage/SidePublication/SidePublication.css
