+++ src/components/LandingPage/SidePublication/SidePublication.data.js
